Added documentation for PUT and code organization

+++ views.py
@@ -90,45 +90,67 @@ class NewsFeed(Resource):
 
 @api.route('/post/<int:post_id>')
 class ReadPost(Resource):
+    @api.response(404, 'Post not found!')
     def get(self, post_id):
         """
         Interact with a specific post
         ---
-        1. Flask-SQLAlchemy looks for the Post with the corresponding ID provided by the client side.
+        1. Flask-SQLAlchemy looks for the Post with the corresponding ID provided by the client side,
+        and it checks if it exists, if it doesn't then it will raise a 404 error.
         2. It gets the first Post it finds with that ID.
         3. Then it gets the Model Schema from 'models.py' so that it can be turned into JSON format.
         4. The Post Schema is then used to dump the data about the Post into JSON and then returns
         a JSON formatted output for Flask-RESTPlus 
         """
         post = Posts.query.filter_by(id=post_id).first()
-        post_schema = PostSchema()
-        output = post_schema.dump(post).data
-        return jsonify({'post': output})
+        if not post:
+            return {'message': 'Post not found!'}, 404
+        else:
+            post_schema = PostSchema()
+            output = post_schema.dump(post).data
+            return jsonify({'post': output})
 
+    @api.response(200, 'Post successfully been updated.')
+    @api.response(404, 'Post not found!')
     @api.expect(user_post)
     def put(self, post_id):
-        # Query the Post
-        post = Posts.query.filter_by(id=post_id)
-        post.update(api.payload)
-        db.session.commit()
-        return {'result': 'Post has been updated'}
+        """
+        Update or Edit a specific post
+        ---
+        1. Flask-SQLAlchemy will query for the post and filters it by provided id, and tries
+        to get the first result of the matched unique id.
+        2. If it the post is not found it will raise a 404 error, else it will
+        update the post with the user provided API Payload, then it commits to the Database.
+        """
+        # Similar to the get method for specific post but updates instead.
+        post = Posts.query.filter_by(id=post_id).first()
+        if not post:
+            return {'message': 'Post not found!'}, 404
+        else:
+            post.update(api.payload)
+            db.session.commit()
+            return {'result': 'Post has been updated'}, 200
 
     @api.response(200, 'Post has successfully been deleted')
+    @api.response(404, 'Post not found!')
     def delete(self, post_id):
         """
         Delete a specific post by id
         ---
         1. Flask-SQLAlchemy queries the Database and filters the result with the ID provided by the
-        client side application.
+        client side application. It first checks if the post exist and raises 404 if not.
         2. Once SQLAlchemy finds that specific post, it is then deleted during the session, then
         commits the changes to the Database.
         3. Then Flask-RESTPlus returns the result
         """
-        # Delete the post from the ID
-        post = Posts.query.filter_by(id=post_id).delete()
-        # Commit those changes
-        db.session.commit()
-        return {'result': 'Post has successfully been deleted'}
+        post = Posts.query.filter_by(id=post_id).first()
+        if not post:
+            return {'message', 'Post not found'}, 404
+        else:
+            post = Posts.query.filter_by(id=post_id).delete()
+            # Commit those changes
+            db.session.commit()
+            return {'result': 'Post has successfully been deleted'}, 200
 
 """ 
 Add Admin Views,
